promote embedWistia to scripts.js
